+++ src/main/java/com/ssafy/petandpeople/application/converter/RedisConverterIfs.java
@@ -1,0 +1,7 @@
+package com.ssafy.petandpeople.application.converter;
+
+public interface RedisConverterIfs<DTO> {
+
+    DTO jsonToDto(String jsonData);
+
+}
