--- src/main/java/com/ssafy/petandpeople/application/dto/EmailDto.java
@@ -1,24 +1,0 @@
-package com.ssafy.petandpeople.application.dto;
-
-public class EmailDto {
-
-    private String email;
-    private String authCode;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getAuthCode() {
-        return authCode;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public void setAuthCode(String authCode) {
-        this.authCode = authCode;
-    }
-
-}
