--- src/main/java/com/ssafy/petandpeople/application/service/PartTimeService.java
@@ -1,31 +1,0 @@
-package com.ssafy.petandpeople.application.service;
-
-import com.ssafy.petandpeople.application.converter.PartTimeConverter;
-import com.ssafy.petandpeople.application.dto.PartTimeDto;
-import com.ssafy.petandpeople.infrastructure.persistence.entity.PartTimeEntity;
-import com.ssafy.petandpeople.infrastructure.persistence.entity.UserEntity;
-import com.ssafy.petandpeople.infrastructure.persistence.repository.PartTimeRepository;
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.stereotype.Service;
-
-@Service
-public class PartTimeService {
-
-    public final PartTimeRepository partTimeRepository;
-    public final UserService userService;
-
-    private PartTimeService(PartTimeRepository partTimeRepository, UserService userService) {
-        this.partTimeRepository = partTimeRepository;
-        this.userService = userService;
-    }
-
-    public Boolean createPartTimePost(PartTimeDto partTimeDto, HttpServletRequest request)  {
-        UserEntity foundUser = userService.findAllByUserKey(request);
-
-        PartTimeEntity partTimeEntity = PartTimeConverter.dtoToEntity(partTimeDto, foundUser);
-        partTimeRepository.save(partTimeEntity);
-
-        return true;
-    }
-
-}
