+++ src/main/java/com/ssafy/petandpeople/common/error/ErrorCodeIfs.java
@@ -6,6 +6,6 @@ public interface ErrorCodeIfs {
 
     Integer getErrorCode();
 
-    String getMessage();
+    String getDescription();
 
 }
