--- src/main/java/com/ssafy/petandpeople/common/exception/email/AuthCodeMismatchException.java
@@ -1,19 +1,0 @@
-package com.ssafy.petandpeople.common.exception.email;
-
-import com.ssafy.petandpeople.common.error.ErrorCodeIfs;
-import com.ssafy.petandpeople.common.error.email.EmailErrorCode;
-import com.ssafy.petandpeople.common.exception.ExceptionIfs;
-
-public class AuthCodeMismatchException extends RuntimeException implements ExceptionIfs {
-
-    private final ErrorCodeIfs errorCodeIfs;
-
-    public AuthCodeMismatchException() {
-        this.errorCodeIfs = EmailErrorCode.AUTH_CODE_MISMATCH;
-    }
-
-    @Override
-    public ErrorCodeIfs getErrorCodeIfs() {
-        return errorCodeIfs;
-    }
-}
