+++ src/main/java/com/ssafy/petandpeople/common/response/Result.java
@@ -1,17 +1,12 @@
-package com.ssafy.petandpeople.presentation.response;
+package com.ssafy.petandpeople.common.response;
 
+import com.ssafy.petandpeople.common.error.ErrorCode;
 import com.ssafy.petandpeople.common.error.ErrorCodeIfs;
 
 public class Result {
-    private static final Result SUCCESS = new Result(200, "성공");
 
     private final Integer resultCode;
     private final String resultMessage;
-
-    private Result(Integer resultCode, String resultMessage) {
-        this.resultCode = resultCode;
-        this.resultMessage = resultMessage;
-    }
 
     private Result(ErrorCodeIfs errorCodeIfs, String resultMessage) {
         this.resultCode = errorCodeIfs.getErrorCode();
@@ -19,15 +14,11 @@ public class Result {
     }
 
     public static Result OK() {
-        return SUCCESS;
+        return new Result(ErrorCode.OK, "성공");
     }
 
     public static Result ERROR(ErrorCodeIfs errorCodeIfs) {
         return new Result(errorCodeIfs, "실패");
-    }
-
-    public static Result ERROR(Integer errorCode) {
-        return new Result(errorCode, "실패");
     }
 
     public Integer getResultCode() {
