--- src/main/java/com/ssafy/petandpeople/infrastructure/persistence/entity/UserEntity.java
@@ -1,119 +1,0 @@
-package com.ssafy.petandpeople.infrastructure.persistence.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "user")
-public class UserEntity {
-
-    public UserEntity() {
-    }
-
-    public UserEntity(String userId, String userPassword, String userName, String userPhoneNumber, String userAddress) {
-        this.userId = userId;
-        this.userPassword = userPassword;
-        this.userName = userName;
-        this.userPhoneNumber = userPhoneNumber;
-        this.userAddress = userAddress;
-    }
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userKey;
-
-    @Column(unique = true, nullable = false)
-    private String userId;
-
-    @Column(nullable = false)
-    private String userPassword;
-
-    @Column(nullable = false)
-    private String userName;
-
-    @Column(unique = true, nullable = false)
-    private String userPhoneNumber;
-
-    @Column(nullable = false)
-    private String userAddress;
-
-    @CreationTimestamp
-    @Column(nullable = false)
-    private LocalDateTime registeredAt;
-
-    @CreationTimestamp
-    @Column(nullable = false)
-    private LocalDateTime lastLoginAt;
-
-    public Long getUserKey() {
-        return userKey;
-    }
-
-    public String getUserId() {
-        return userId;
-    }
-
-    public String getUserPassword() {
-        return userPassword;
-    }
-
-    public String getUserName() {
-        return userName;
-    }
-
-    public String getUserPhoneNumber() {
-        return userPhoneNumber;
-    }
-
-    public String getUserAddress() {
-        return userAddress;
-    }
-
-    public LocalDateTime getRegisteredAt() {
-        return registeredAt;
-    }
-
-    public LocalDateTime getLastLoginAt() {
-        return lastLoginAt;
-    }
-
-    public void setUserKey(Long userKey) {
-        this.userKey = userKey;
-    }
-
-    public void setUserId(String userId) {
-        this.userId = userId;
-    }
-
-    public void setUserPassword(String userPassword) {
-        this.userPassword = userPassword;
-    }
-
-    public void setUserName(String userName) {
-        this.userName = userName;
-    }
-
-    public void setUserPhoneNumber(String userPhoneNumber) {
-        this.userPhoneNumber = userPhoneNumber;
-    }
-
-    public void setUserAddress(String userAddress) {
-        this.userAddress = userAddress;
-    }
-
-    public void setRegisteredAt(LocalDateTime registeredAt) {
-        this.registeredAt = registeredAt;
-    }
-
-    public void setLastLoginAt(LocalDateTime lastLoginAt) {
-        this.lastLoginAt = lastLoginAt;
-    }
-
-}
