--- src/main/java/com/ssafy/petandpeople/infrastructure/persistence/repository/PartTimeRepository.java
@@ -1,10 +1,0 @@
-package com.ssafy.petandpeople.infrastructure.persistence.repository;
-
-import com.ssafy.petandpeople.infrastructure.persistence.entity.PartTimeEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PartTimeRepository extends JpaRepository<PartTimeEntity, Long> {
-
-}
