--- src/main/java/com/ssafy/petandpeople/infrastructure/persistence/repository/UserSecurityRepository.java
@@ -1,10 +1,0 @@
-package com.ssafy.petandpeople.infrastructure.persistence.repository;
-
-import com.ssafy.petandpeople.infrastructure.persistence.entity.UserSecurityEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserSecurityRepository extends JpaRepository<UserSecurityEntity, String> {
-
-}
