--- src/main/resources/application-test.yaml
@@ -1,12 +1,0 @@
-spring:
-  config:
-    import: optional:file:.env[.properties]
-  datasource:
-    url: ${TEST_DB_URL}
-    username: ${TEST_DB_USERNAME}
-    password: ${TEST_DB_PASSWORD}
-    driver-class-name: ${TEST_DB_DRIVER_CLASS_NAME}
-  jpa:
-    hibernate:
-      ddl-auto: ${TEST_JPA_HIBERNATE_DDL_AUTO}
-    show-sql: ${TEST_JPA_SHOW_SQL}
