+++ src/main/resources/application.properties
@@ -1,0 +1,9 @@
+spring.application.name=petandpeople
+
+spring.datasource.url=jdbc:mysql://localhost:3306/test
+spring.datasource.username=test
+spring.datasource.password=test
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.hikari.maximum-pool-size=10
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
