--- src/main/resources/application.yaml
@@ -1,30 +1,0 @@
-spring:
-  config:
-    import: optional:file:.env[.properties]
-  mail:
-    host: ${MAIL_HOST}
-    port: ${MAIL_PORT}
-    username: ${MAIL_USERNAME}
-    password: ${MAIL_PASSWORD}
-    properties:
-      mail.smtp.auth: ${MAIL_SMTP_AUTH}
-      mail.smtp.ssl.enable: ${MAIL_SMTP_SSL_ENABLE}
-      mail.smtp.timeout: ${MAIL_SMTP_TIMEOUT}
-
-  datasource:
-    url: ${DB_URL}
-    username: ${DB_USERNAME}
-    password: ${DB_PASSWORD}
-    driver-class-name: ${DB_DRIVER_CLASS_NAME}
-    hikari:
-      maximum-pool-size: ${DB_HIKARI_MAX_POOL_SIZE}
-
-  jpa:
-    hibernate:
-      ddl-auto: ${JPA_HIBERNATE_DDL_AUTO}
-    show-sql: ${JPA_SHOW_SQL}
-    database-platform: ${JPA_DATABASE_PLATFORM}
-
-server:
-  servlet:
-    context-path: ${SERVER_CONTEXT_PATH}
